requirements.txt
ansible==1.9.2
asgi-redis==0.13.1
cairosvg==1.0.19
channels==1.0.2
daphne==1.0.1
Django==1.11
django-extensions==1.6.7
django-filter==1.0.2
django-global-permissions==0.2.2
django-sendfile==0.3.11
django-webpack-loader==0.3.0
djangoql==0.8.2
djangorestframework==3.3.3
graphene==1.1.3
graphene-django==1.2.1
graphql-core==1.0.1
graphql-relay==0.4.5
Jinja2==2.8
lxml==3.4.4
psycopg2==2.6.1
pycrypto==2.6.1
Pillow==2.9.0
redis==2.10.5
requests==2.7.0
selenium==3.3.1
Twisted==16.6.0
